+++ packages/core/src/index.ts
@@ -43,7 +43,7 @@ export { LogMediator, LogFilter, MsgLogFilter } from './services/log-mediator';
  export { AppInitializer } from './services/app-initializer';
  export { ExtensionsManager } from './services/extensions-manager';
  export { ModuleManager } from './services/module-manager';
- export { ExtensionItem1, ExtensionItem2 } from './utils/get-extension-provider';
+ export { ExtensionOptions } from './utils/get-extension-provider';
  export { ExtensionsContext } from './services/extensions-context';
  export { MetadataPerMod1, MetadataPerMod2 } from './types/metadata-per-mod';
  export { ControllersMetadata1, ControllersMetadata2 } from './types/controller-metadata';